esp_demos: cifar10 add jpg predict

+++ examples/dl/cifar10/esp_cifar10/main/model/include/predict.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 /**
  * @brief cifar10 model init
  * 
@@ -12,4 +14,4 @@ void cifar10_model_init();
  * @param pic 
  * @param size 
  */
-void cifar10_model_predict(const float *pic, int size);
+void cifar10_model_predict(const uint8_t *pic, int size);
